requirements.txt
fastapi
uvicorn
python-multipart
ffmpeg-python
faster-whisper
google-generativeai
